update info button working
SIGA SIGA! PARA CIMA DELES COMO D. AFONSO HENRIQUES!

+++ classes/product.class.php
@@ -5,7 +5,7 @@ declare(strict_types = 1);
 class Product{
 
     private int $id;
-    private int $price;
+    private float $price;
     private int $quantity;
     private string $name;
     private string $description;
@@ -13,7 +13,7 @@ class Product{
     private int $category;
     private int $brand;
 
-    public function __construct(int $id, int $price, int $quantity, string $name, string $description, int $owner, int $category, int $brand){
+    public function __construct(int $id, float $price, int $quantity, string $name, string $description, int $owner, int $category, int $brand){
         $this->id = $id;
         $this->price = $price;
         $this->quantity = $quantity;
@@ -28,7 +28,7 @@ class Product{
         return $this->id;
     }
 
-    public function getPrice() : int{
+    public function getPrice() : float{
         return $this->price;
     }
 
@@ -68,7 +68,7 @@ class Product{
 
         return new Product(
             intval($product['id']),
-            intval($product['price']),
+            floatval($product['price']),
             intval($product['quantity']),
             $product['name'],
             $product['description'],
@@ -78,7 +78,7 @@ class Product{
         );
     }
 
-    static public function getProductByPrice(PDO $db, int $lowerPrice, int $upperPrice){
+    static public function getProductByPrice(PDO $db, float $lowerPrice, float $upperPrice){
         $stmt = $db->prepare(
             'SELECT id, price, quantity, name, description, owner, category, brand
             FROM PRODUCTS
@@ -89,7 +89,7 @@ class Product{
         while($productDB = $stmt->fetch()){
             $product = new Product(
                 intval($productDB['id']),
-                intval($productDB['price']),
+                floatval($productDB['price']),
                 intval($productDB['quantity']),
                 $productDB['name'],
                 $productDB['description'],
@@ -113,7 +113,7 @@ class Product{
         while($productDB = $stmt->fetch()){
             $product = new Product(
                 intval($productDB['id']),
-                intval($productDB['price']),
+                floatval($productDB['price']),
                 intval($productDB['quantity']),
                 $productDB['name'],
                 $productDB['description'],
@@ -136,7 +136,7 @@ class Product{
         while($productDB = $stmt->fetch()){
             $product = new Product(
                 intval($productDB['id']),
-                intval($productDB['price']),
+                floatval($productDB['price']),
                 intval($productDB['quantity']),
                 $productDB['name'],
                 $productDB['description'],
